pyproject.toml
[project]
name = "dzd"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "datasets>=4.3.0",
    "fastparquet>=2024.11.0",
    "h5py>=3.15.1",
    "ipykernel>=6.30.1",
    "jupyter>=1.1.1",
    "kaggle>=1.7.4.5",
    "locust>=2.42.3",
    "matplotlib>=3.10.6",
    "mkdocs-material>=9.6.21",
    "mlserver>=1.7.1",
    "mlserver-sklearn>=1.7.1",
    "mlserver-xgboost>=1.7.1",
    "mrjob>=0.7.4",
    "nbconvert[webpdf]>=7.16.6",
    "pandas>=2.3.3",
    "pyarrow>=21.0.0",
    "scikit-learn>=1.7.2",
    "tqdm>=4.67.1",
    "uvloop==0.21.0",
    "xgboost>=3.1.1",
]
